rename httpthread files to not conflict with their non threaded counterpart
this is a workaround for a build artifact of certain toolchains that are storing some cached files in the same namespace, regardless of path
