+++ Builds/VisualStudio/stellar-core.vcxproj
@@ -430,6 +430,10 @@ exit /b 0
     </PreLinkEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="..\..\lib\httpthreaded\connection_mt.cpp" />
+    <ClCompile Include="..\..\lib\httpthreaded\reply_mt.cpp" />
+    <ClCompile Include="..\..\lib\httpthreaded\request_parser_mt.cpp" />
+    <ClCompile Include="..\..\lib\httpthreaded\server_mt.cpp" />
     <ClCompile Include="..\..\lib\spdlog.cpp" />
     <ClCompile Include="..\..\lib\tracy\public\TracyClient.cpp" />
     <ClCompile Include="..\..\lib\util\siphash.cpp" />
@@ -592,6 +596,7 @@ exit /b 0
     <ClCompile Include="..\..\src\ledger\SorobanMetrics.cpp" />
     <ClCompile Include="..\..\src\ledger\TrustLineWrapper.cpp" />
     <ClCompile Include="..\..\src\main\Diagnostics.cpp" />
+    <ClCompile Include="..\..\src\main\QueryServer.cpp" />
     <ClCompile Include="..\..\src\main\SettingsUpgradeUtils.cpp" />
     <ClCompile Include="..\..\src\main\test\ApplicationUtilsTests.cpp" />
     <ClCompile Include="..\..\src\main\test\CommandHandlerTests.cpp" />
@@ -657,6 +662,7 @@ exit /b 0
     <ClCompile Include="..\..\src\transactions\ManageOfferOpFrameBase.cpp" />
     <ClCompile Include="..\..\src\transactions\ManageSellOfferOpFrame.cpp" />
     <ClCompile Include="..\..\src\transactions\MergeOpFrame.cpp" />
+    <ClCompile Include="..\..\src\transactions\MutableTransactionResult.cpp" />
     <ClCompile Include="..\..\src\transactions\OfferExchange.cpp" />
     <ClCompile Include="..\..\src\transactions\OperationFrame.cpp" />
     <ClCompile Include="..\..\src\transactions\PathPaymentOpFrameBase.cpp" />
@@ -699,6 +705,7 @@ exit /b 0
     <ClCompile Include="..\..\src\transactions\test\SignatureUtilsTest.cpp" />
     <ClCompile Include="..\..\src\transactions\test\SorobanTxTestUtils.cpp" />
     <ClCompile Include="..\..\src\transactions\test\SponsorshipTestUtils.cpp" />
+    <ClCompile Include="..\..\src\transactions\test\TransactionTestFrame.cpp" />
     <ClCompile Include="..\..\src\transactions\test\TxEnvelopeTests.cpp" />
     <ClCompile Include="..\..\src\transactions\test\TxResultsTests.cpp" />
     <ClCompile Include="..\..\src\transactions\TransactionBridge.cpp" />
@@ -716,6 +723,7 @@ exit /b 0
     <ClCompile Include="..\..\src\util\LogSlowExecution.cpp" />
     <ClCompile Include="..\..\src\util\RandHasher.cpp" />
     <ClCompile Include="..\..\src\util\Scheduler.cpp" />
+    <ClCompile Include="..\..\src\util\test\BinaryFuseTests.cpp" />
     <ClCompile Include="..\..\src\util\test\MetricTests.cpp" />
     <ClCompile Include="..\..\src\util\test\SchedulerTests.cpp" />
     <ClCompile Include="..\..\src\util\TxResource.cpp" />
@@ -877,6 +885,12 @@ exit /b 0
     <ClInclude Include="..\..\lib\binaryfusefilter.h" />
     <ClInclude Include="..\..\lib\catch.hpp" />
     <ClInclude Include="..\..\lib\clara.hpp" />
+    <ClInclude Include="..\..\lib\httpthreaded\connection_mt.hpp" />
+    <ClInclude Include="..\..\lib\httpthreaded\header_mt.hpp" />
+    <ClInclude Include="..\..\lib\httpthreaded\reply_mt.hpp" />
+    <ClInclude Include="..\..\lib\httpthreaded\request_mt.hpp" />
+    <ClInclude Include="..\..\lib\httpthreaded\request_parser_mt.hpp" />
+    <ClInclude Include="..\..\lib\httpthreaded\server_mt.hpp" />
     <ClInclude Include="..\..\lib\util\finally.h" />
     <ClInclude Include="..\..\lib\fmt\include\fmt\format.h" />
     <ClInclude Include="..\..\lib\util\siphash.h" />
@@ -1015,6 +1029,7 @@ exit /b 0
     <ClInclude Include="..\..\src\ledger\SorobanMetrics.h" />
     <ClInclude Include="..\..\src\ledger\TrustLineWrapper.h" />
     <ClInclude Include="..\..\src\main\Diagnostics.h" />
+    <ClInclude Include="..\..\src\main\QueryServer.h" />
     <ClInclude Include="..\..\src\main\SettingsUpgradeUtils.h" />
     <ClInclude Include="..\..\src\overlay\BanManager.h" />
     <ClInclude Include="..\..\src\overlay\BanManagerImpl.h" />
@@ -1068,6 +1083,7 @@ exit /b 0
     <ClInclude Include="..\..\src\transactions\ManageOfferOpFrameBase.h" />
     <ClInclude Include="..\..\src\transactions\ManageSellOfferOpFrame.h" />
     <ClInclude Include="..\..\src\transactions\MergeOpFrame.h" />
+    <ClInclude Include="..\..\src\transactions\MutableTransactionResult.h" />
     <ClInclude Include="..\..\src\transactions\OfferExchange.h" />
     <ClInclude Include="..\..\src\transactions\OperationFrame.h" />
     <ClInclude Include="..\..\src\transactions\PathPaymentOpFrameBase.h" />
@@ -1083,6 +1099,7 @@ exit /b 0
     <ClInclude Include="..\..\src\transactions\SponsorshipUtils.h" />
     <ClInclude Include="..\..\src\transactions\test\SorobanTxTestUtils.h" />
     <ClInclude Include="..\..\src\transactions\test\SponsorshipTestUtils.h" />
+    <ClInclude Include="..\..\src\transactions\test\TransactionTestFrame.h" />
     <ClInclude Include="..\..\src\transactions\TransactionBridge.h" />
     <ClInclude Include="..\..\src\transactions\TransactionFrame.h" />
     <ClInclude Include="..\..\src\transactions\TransactionFrameBase.h" />
