+++ lib/httpthreaded/reply_mt.hpp
@@ -14,7 +14,7 @@
 // else.
 #include "util/asio.h"
 
-#include "header.hpp"
+#include "header_mt.hpp"
 #include <string>
 #include <vector>
