+++ lib/httpthreaded/server_mt.cpp
@@ -7,7 +7,7 @@
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 //
-#include "server.hpp"
+#include "server_mt.hpp"
 
 #include "util/GlobalChecks.h"
 #include "util/Thread.h"
